+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2025-12-26
+
+- Added support for customizing the refresh interval (`ag-usage.refreshInterval`)
+- The timer is now visible only when the reset quota time is triggered
+- Parallelized port discovery for faster connection
+
 ## [1.0.0] - 2025-12-24
 
 - Initial release